requirements.txt
flask==2.3.3
google-generativeai==0.3.0
python-dotenv==0.19.2
